Confirmation Modal Added
Added 'react-modal' dependency.

Added a confirmation modal that prompts upon removing an expense.

Removed the RemoveExpense test for now as the portal functionality with react-modal is poorly documented. TBD

Lots of small CSS fixes.

Added styling for the modal and associated elements.

Added a color bar to the top of the page :D.