Large patch for pre-release of email auth
Updated color bar, logout button, and secondary button styles.

Checked on Github issue regarding filters.

Created login/signup pages.

Fixed an error associated with a Modal from 'react-modal' not being aria-accessible due to lack of DOM attachment.

Added 'formik' and 'yup' as dependencies for the sign-in and signup forms.

Updated syntax for imports in some files (to be done: update the syntax for imports in all files).

README update.

Updates to the login page to prepare it for production.

Disabled the login via email and create account inputs for now.

Added styling for disabled elements.